+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+#### itachallenge-challenge-1.0.0-RELEASE
+* First version
+
+#### docker-compose-1.0.3
+* MongoDB Challenge Docker Volumes parameterized
+
 #### docker-compose-1.0.2
 * Docker image Challenge (version)
 * MongoDB Access Port (from another Docker container)